Django/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
pytz
sqlparse